+++ anno-docimal-ast/src/main/java/com/blackbuild/annodocimal/ast/InlineJavadocsVisitor.java
@@ -75,7 +75,8 @@ public class InlineJavadocsVisitor extends ClassCodeVisitorSupport {
     }
 
     private void addJavadocAsAnnotation(AnnotatedNode node) {
-        if (!node.getAnnotations(AnnoDocUtil.ANNODOC_ANNOTATION).isEmpty()) return;
+        if (!node.getAnnotations(AnnoDocUtil.ANNODOC_ANNOTATION).isEmpty()
+                || AnnoDocUtil.hasRuntimeGroovydoc(node)) return;
 
         String javadoc = sourceExtractor.getJavaDoc(node);
         if (javadoc != null && !javadoc.isBlank())